refactor(boot): selfrepair + systemtasks extractions, HTTP-surface file split (Phase 2.6 stages 1-2)
kernel/selfrepair: the 3,226-line auto-repair coordinator leaves
cmd/agezt (imports were already kernel-clean); WireAutoRepair+Mailbox
exported, rest package-private. kernel/cadence/systemtasks: the 7
system-task executors join their catalogue (which was already in
cadence) — package doc pins the import-cycle constraint. main.go's
HTTP surfaces move to httpsurfaces.go (same package: no export churn,
no env-guard blindness) with two cleanups: the duplicated
tenant-resolver closure deduped into tenantAPIEngine, buildRESTAPI's
60-line SetMetrics closure lifted into named restMetrics.
kernelAPIEngine → api_engine.go (deliberately NOT kernel/restapi —
that package's interface-only posture is the point). Status setters
hoisted above srv.Start. config_inventory gains both new roots.
main.go 4,885 → 3,810.
