+++ kernel/selfrepair/coverage_test.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 
-package main
+package selfrepair
 
 import (
 	"strings"
