+++ kernel/selfrepair/selfrepair_test.go
@@ -1,6 +1,6 @@
 // SPDX-License-Identifier: MIT
 
-package main
+package selfrepair
 
 import (
 	"context"
